update and split file
