+++ package.json
@@ -24,6 +24,7 @@
     "@tanstack/react-table": "^8.13.2",
     "axios": "^1.6.7",
     "dayjs": "^1.11.10",
+    "downshift": "^8.4.0",
     "localforage": "^1.10.0",
     "lodash": "^4.17.21",
     "match-sorter": "^6.3.4",
