+++ src/base/PageLayout/index.tsx
@@ -4,13 +4,13 @@ import { Outlet } from 'react-router';
 import Header from './Header';
 import { CONTAINER_PADDING_X, HEADER_HEIGHT } from '@base/config/constants';
 import { useMatch } from 'react-router-dom';
-import CreateProjectModal from '@pages/HomePage/Components/CreateProjectModal';
+import CreateProjectModal from '@pages/HomePage/components/CreateProject/CreateProjectModal';
 
 const PageLayout = () => {
   const isHome = useMatch('/');
   const isLogin = useMatch('/login');
 
-  // const [isOpen, setIsOpen] = useState(true);
+  const [isOpen, setIsOpen] = useState(true);
   return (
     <Stack>
       {!isLogin && <Header />}
@@ -22,7 +22,7 @@ const PageLayout = () => {
       >
         <Outlet />
       </Stack>
-      {/* <CreateProjectModal isOpen={isOpen} title='' onClose={() => setIsOpen(false)} size='lg' /> */}
+      <CreateProjectModal isOpen={isOpen} title='' onClose={() => setIsOpen(false)} size='lg' />
     </Stack>
   );
 };
